+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Moved app views from project to new xtc `views` dir. Patched Express to search multiple views dirs.
 - Moved xtc app routes from project to xtc controllers.
 - xtc app routes are now mounted at `/xtc`.
+- You can specify the port that xtc listens to by calling `xtc start` with `-p [number]` or `--port [number]`.
 
 ## 0.8.0-beta7
 - xtc-generator is now part of the xtc package and can be run from the project folder.