+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Moved the Handlebars helper Hipsum.js to project dependencies.
 - The Handlebars instance is passed to project. Project template helpers are now registered at the project level instead of in the xtc module.
 - Added a new project-level helpers module to add new or override default helper functions at `app.xtc.helpers`.
+- Ability to add custom grunt plugins at project level.
 
 ## 0.8.0-beta7
 - xtc-generator is now part of the xtc package and can be run from the project folder.